log: Create abstract log implementation interface
New abstraction layer for logging implementation and create
separate sylog logging implemntation.
