+++ lib/log/Makefile
@@ -34,7 +34,7 @@
 SPDK_ROOT_DIR := $(abspath $(CURDIR)/../..)
 include $(SPDK_ROOT_DIR)/mk/spdk.common.mk
 
-C_SRCS = log.c
+C_SRCS = log.c log_syslog.c
 LIBNAME = log
 
 DIRS-y = rpc
