+++ test/lib/log/log_ut.c
@@ -40,6 +40,7 @@
 #include "spdk/log.h"
 
 #include "log.c"
+#include "log_syslog.c"
 
 static void
 log_test(void)
