+++ test/lib/scsi/dev/dev_ut.c
@@ -38,6 +38,7 @@
 #include "CUnit/Basic.h"
 #include "spdk_cunit.h"
 
+#include "log/log_syslog.c"
 #include "dev.c"
 #include "port.c"
