+++ test/lib/scsi/lun/lun_ut.c
@@ -38,6 +38,7 @@
 
 #include "spdk_cunit.h"
 
+#include "log/log_syslog.c"
 #include "task.c"
 #include "lun.c"
 #include "lun_db.c"
