+++ test/lib/scsi/scsi_bdev/Makefile
@@ -39,6 +39,7 @@ SPDK_LIB_LIST = log util cunit
 
 CFLAGS += -I$(SPDK_ROOT_DIR)/test
 CFLAGS += -I$(SPDK_ROOT_DIR)/lib/scsi
+CFLAGS += -I$(SPDK_ROOT_DIR)/lib
 LIBS += $(SPDK_LIB_LINKER_ARGS)
 LIBS += -lcunit
