+++ test/lib/scsi/scsi_bdev/scsi_bdev_ut.c
@@ -35,6 +35,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "log/log_syslog.c"
 #include "task.c"
 #include "scsi_bdev.c"
