+++ test/lib/test_env.c
@@ -37,6 +37,7 @@
 #include <stdint.h>
 
 #include "spdk/env.h"
+#include "log/log_syslog.c"
 
 void *
 spdk_dma_malloc(size_t size, size_t align, uint64_t *phys_addr)
