+++ test/lib/util/io_channel/io_channel_ut.c
@@ -37,6 +37,7 @@
 
 #include "spdk_cunit.h"
 
+#include "log/log_syslog.c"
 #include "util/io_channel.c"
 
 static void
